feat:  horizontal scaling for app, and refactor;
