+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "start:dev": "ts-node-dev src/server.ts",
     "build": "tsc",
-    "start:prod": "npm run build && node dist/server.js"
+    "start:prod": "npm run build && node dist/server.js",
+    "start:multi": "ts-node src/cluster.ts"
   },
   "dependencies": {
     "uuid": "^9.0.0"
